refactor: extract window commands into their own file, move active into element.ts
Window-related commands (setWindow, getWindowHandle(s), getWindowSize/Rect,
setWindowRect, maximize/minimize/fullScreen, closeWindow) were scattered
across context.ts, general.ts, and navigation.ts - pulled them into a new
window.ts to match this codebase's one-file-per-domain convention.

active() returns an element, so it belongs with the other element commands
rather than in the general.ts grab-bag.
